pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autonode"
version = "0.1.0"
dependencies = [
    "python-dotenv",
    "pyyaml>=6.0",
    "fastapi",
    "langchain>=0.3.0",
    "langchain-openai",
    "langchain-core",
    "langchain-community",
    "langgraph>=0.2",
    "langgraph-checkpoint-sqlite",
    "langsmith>=0.1",
    "aider-chat",
    "docker>=7.0.0",
    "uvicorn",
]

[project.optional-dependencies]
dev = [
    "types-PyYAML",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "black>=24.0.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
mypy_path = ["src"]
explicit_package_bases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[project.scripts]
autonode = "autonode.presentation.cli:main"
lint = "scripts.dev:lint"
fix = "scripts.dev:fix"

[tool.setuptools]
packages = ["autonode", "scripts"]

[tool.setuptools.package-dir]
"autonode" = "src/autonode"
"scripts" = "scripts"